[FXIOS-10255] Minor refactor, move spam filtering into a separate class and protocol
